+++ index.d.ts
@@ -1,0 +1,9 @@
+export {
+  MediaQueries,
+  Orientation,
+  Props,
+  up,
+  down,
+  between,
+  only,
+} from './core';